update 7
Finished the basic functionality

+++ app/src/main/java/com/example/atrack/data/Subject.kt
@@ -11,5 +11,5 @@ data class Subject(
     var subCode: String,
     var nPresent: Int,
     var nTotal: Int,
-    var percent: Int
+    var percent: Float
 )
